# src/test/resources/shiptime_mock_api/openapi/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "SANDBOX REST API",
    "description": "Welcome to the Rest API documentation. Don't have API keys? Login or create an account <a href=\"https://restapi.appspaces.ca/\">here</a> to get started.",
    "version": "0.1"
  },
  "servers": [
    {
      "url": "http://restapi.appspaces.ca/rest"
    }
  ],
...
  "paths": {
...
    "/shipments": {
      "get": {
        "tags": [
          "shipments"
        ],
        "summary": "Get all shipments of the customer",
        "operationId": "shipments-get-all",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number. Default is 0",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "Number of shipments to retrieve. Default is 10",
            "schema": {
              "type": "integer",
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Shipments Retrieved Successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad request. The request was invalid because the body was missing a required parameter and/or the parameters did not conform to data validation. See error messages in body.",
            "content": {}
          },
          "401": {
            "description": "Unauthorized. Client authentication failed. (Invalid username/password)",
            "content": {}
          },
          "404": {
            "description": "Resource not found for the ID",
            "content": {}
          }
        }
      }
    },
    "/shipments/": {
      "post": {
        "tags": [
          "shipments"
        ],
        "summary": "Create a shipment. The operation will result in billing the account.",
        "operationId": "shipments-post",
        "requestBody": {
          "description": "The shipment request to process",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShipRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Output type",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShipResponse"
                }
              }
            }
          },
          "201": {
            "description": "Shipment processed successfully",
            "content": {}
          },
          "400": {
            "description": "Bad request. The request was invalid because the body was missing a required parameter and/or the parameters did not conform to data validation. See error messages in body.",
            "content": {}
          },
          "401": {
            "description": "Unauthorized. Client authentication failed. (Invalid username/password)",
            "content": {}
          },
          "500": {
            "description": "A server-side error has occured.",
            "content": {}
          }
        },
        "x-codegen-request-body-name": "body"
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PagedResponse": {
        "type": "object",
        "properties": {
          "content": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {}
            }
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "last": {
            "type": "boolean"
          }
        }
      },
      "LineItemModel": {
        "required": [
          "height",
          "length",
          "weight",
          "width"
        ],
        "type": "object",
        "properties": {
          "length": {
            "type": "number",
            "description": "Length of the package (IMPERIAL: inches/METRIC: centimetres). If a decimal number is declared, please limit the fractional part to two digits"
          },
          "width": {
            "type": "number",
            "description": "Width of the package (IMPERIAL: inches/METRIC: centimetres). If a decimal number is declared, please limit the fractional part to two digits"
          },
          "height": {
            "type": "number",
            "description": "Height of the package (IMPERIAL: inches/METRIC: centimetres). If a decimal number is declared, please limit the fractional part to two digits"
          },
          "weight": {
            "type": "number",
            "description": "Weight of the package (IMPERIAL: pounds/METRIC: kilograms). If a decimal number is declared, please limit the fractional part to three digits"
          },
          "declaredValue": {
            "$ref": "#/components/schemas/MoneyAmountModel"
          },
          "description": {
            "type": "string",
            "description": "Description of the item"
          },
          "nmfcCode": {
            "type": "string",
            "description": "National Motor Freight Classification Code of the item, Only for PALLET Packaging type"
          },
          "freightClass": {
            "type": "string",
            "description": "Required only for International PALLET Packaging type."
          }
        },
        "description": "Represents an item in a shipment"
      },
      "MoneyAmountModel": {
        "required": [
          "amount",
          "currency"
        ],
        "type": "object",
        "properties": {
          "currency": {
            "type": "string",
            "description": "Currency (USD or CAD)",
            "enum": [
              "CAD",
              "USD"
            ]
          },
          "amount": {
            "type": "integer",
            "description": "Money amount in cents",
            "format": "int32"
          }
        },
        "description": "An amount and a currency"
      },
      "CustomsInvoiceModel": {
        "required": [
          "dutiesAndTaxes"
        ],
        "type": "object",
        "properties": {
          "invoiceContact": {
            "$ref": "#/components/schemas/InvoiceContactModel"
          },
          "dutiesAndTaxes": {
            "$ref": "#/components/schemas/DutiesAndTaxesModel"
          },
          "currency": {
            "type": "string",
            "description": "Currency of this invoice (USD or CAD)",
            "enum": [
              "CAD",
              "USD"
            ]
          },
          "invoiceItems": {
            "type": "array",
            "description": "List of Invoice Items",
            "items": {
              "$ref": "#/components/schemas/InvoiceItemModel"
            }
          },
          "reasonForExport": {
            "type": "string",
            "description": "Reason For Export",
            "enum": [
              "COMMERCIAL",
              "GIFT",
              "SAMPLE",
              "PERSONAL",
              "RETURN_AND_REPAIR"
            ]
          }
        },
        "description": "Customs Invoice"
      },
      "PickupDetailModel": {
        "required": [
          "closeTime",
          "location",
          "pickupDate",
          "readyTime"
        ],
        "type": "object",
        "properties": {
          "location": {
            "type": "string",
            "description": "Location of pickup",
            "enum": [
              "FrontDoor",
              "Office",
              "Lobby",
              "MailRoom",
              "Mailbox",
              "PartsDepartment",
              "Pharmacy",
              "Receiving",
              "Reception",
              "ServiceCounter",
              "SideDoor",
              "BackDoor",
              "Basement",
              "Garage",
              "Shipping",
              "Kiosk",
              "LoadingDock",
              "Warehouse",
              "Other"
            ]
          },
          "otherLocation": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "Other location of pickup"
          },
          "pickupTip": {
            "$ref": "#/components/schemas/MoneyAmountModel"
          },
          "pickupDate": {
            "type": "string",
            "description": "The date of the pickup. Has to be on a business day in most cases. (Format: \"yyyy-MM-dd\")",
            "format": "date"
          },
          "readyTime": {
            "type": "string",
            "description": "The time the shipments will be ready for pickup in 24 hour format. (Format: \"HH:mm\")"
          },
          "closeTime": {
            "type": "string",
            "description": "Latest time the pickup can be done in 24 hour format.\nCannot be later than pickup cutoff time provided along with the rate quote.\nNormally there must be few hours interval between readyTime and closeTime. (Format: \"HH:mm\")"
          }
        },
        "description": "Represents pickup information either in a request or in response. The pickup address will be the origin address of the shipment."
      },
      "DutiesAndTaxesModel": {
        "required": [
          "dutiable"
        ],
        "type": "object",
        "properties": {
          "dutiable": {
            "type": "boolean",
            "description": "Is the shipment dutiable? Non-dutiable are shipments with no monetary value and are known as documents or general correspondence. All other shipments classified as dutiable shipments by Customs and may be levied customs duties and taxes for entrance into the destination country."
          },
          "paidBy": {
            "type": "string",
            "description": "Who pays duties and taxes? (Consignee or Shipper)",
            "enum": [
              "CONSIGNEE",
              "SHIPPER"
            ]
          }
        }
      },
      "AddressModel": {
        "required": [
          "attention",
          "city",
          "companyName",
          "countryCode",
          "phone",
          "postalCode",
          "state",
          "streetAddress"
        ],
        "type": "object",
        "properties": {
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "Company name. Maximum: 40 characters."
          },
          "streetAddress": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Street Address. Maximum: 50 characters."
          },
          "streetAddress2": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "description": "Street Address Line 2. Maximum: 35 characters."
          },
          "city": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "City. Maximum: 40 characters."
          },
          "countryCode": {
            "type": "string",
            "description": "Country Code"
          },
          "state": {
            "type": "string",
            "description": "State or Province. Required for CA and US"
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "description": "Postal Code or ZIP code. Maximum: 10 characters."
          },
          "attention": {
            "maxLength": 40,
            "minLength": 2,
            "type": "string",
            "description": "Recipient's Name. Minimum: 2 characters. Maximum: 40 characters."
          },
          "email": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "Email Address. Maximum: 40 characters."
          },
          "phone": {
            "maxLength": 15,
            "minLength": 10,
            "type": "string",
            "description": "Phone Number. Minimum: 10 characters. Maximum: 15 characters. For Canada and USA the format should be NNN NNN NNNN, for other countries adhere to the E.123 international notation. Example, +22 607 123 4567"
          },
          "instructions": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "description": "Special Instructions. Maximum: 60 characters."
          },
          "residential": {
            "type": "boolean",
            "description": "Residential Address?"
          },
          "notify": {
            "type": "boolean",
            "description": "Notify Recipient?"
          }
        },
        "description": "Address Info"
      },
      "InvoiceContactModel": {
        "required": [
          "attention",
          "city",
          "companyName",
          "countryCode",
          "phone",
          "postalCode",
          "state",
          "streetAddress"
        ],
        "type": "object",
        "properties": {
          "companyName": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "Company name. Maximum: 40 characters."
          },
          "streetAddress": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Street Address. Maximum: 50 characters."
          },
          "streetAddress2": {
            "maxLength": 35,
            "minLength": 0,
            "type": "string",
            "description": "Street Address Line 2. Maximum: 35 characters."
          },
          "city": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "City. Maximum: 40 characters."
          },
          "countryCode": {
            "type": "string",
            "description": "Country Code"
          },
          "state": {
            "type": "string",
            "description": "State or Province. Required for CA and US"
          },
          "postalCode": {
            "maxLength": 10,
            "minLength": 0,
            "type": "string",
            "description": "Postal Code or ZIP code. Maximum: 10 characters."
          },
          "attention": {
            "maxLength": 40,
            "minLength": 2,
            "type": "string",
            "description": "Recipient's Name. Minimum: 2 characters. Maximum: 40 characters."
          },
          "email": {
            "maxLength": 40,
            "minLength": 0,
            "type": "string",
            "description": "Email Address. Maximum: 40 characters."
          },
          "phone": {
            "maxLength": 15,
            "minLength": 10,
            "type": "string",
            "description": "Phone Number. Minimum: 10 characters. Maximum: 15 characters. For Canada and USA the format should be NNN NNN NNNN, for other countries adhere to the E.123 international notation. Example, +22 607 123 4567"
          },
          "instructions": {
            "maxLength": 60,
            "minLength": 0,
            "type": "string",
            "description": "Special Instructions. Maximum: 60 characters."
          },
          "residential": {
            "type": "boolean",
            "description": "Residential Address?"
          },
          "notify": {
            "type": "boolean",
            "description": "Notify Recipient?"
          },
          "customsBroker": {
            "type": "string",
            "description": "Customs Broker"
          },
          "shipperTaxId": {
            "type": "string",
            "description": "Shipper's Tax Id"
          }
        },
        "description": "Address Info"
      },
      "RateRequest": {
        "required": [
          "from",
          "lineItems",
          "packageType",
          "shipDate",
          "to",
          "unitOfMeasurement"
        ],
        "type": "object",
        "properties": {
          "from": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "to": {
            "$ref": "#/components/schemas/AddressModel"
          },
          "packageType": {
            "type": "string",
            "description": "Packaging type",
            "enum": [
              "PACKAGE",
              "LETTER",
              "PAK",
              "MY_ENVELOPE",
              "PALLET"
            ]
          },
          "lineItems": {
            "type": "array",
            "description": "Collection of shipment items",
            "items": {
              "$ref": "#/components/schemas/LineItemModel"
            }
          },
          "unitOfMeasurement": {
            "type": "string",
            "description": "Units of measurements of package dimensions and weight: inches and pounds for IMPERIAL, and centimeters and kilograms for METRIC",
            "enum": [
              "IMPERIAL",
              "METRIC"
            ]
          },
          "serviceOptions": {
            "type": "array",
            "description": "Service options",
            "items": {
              "type": "string",
              "enum": [
                "APPOINTMENT",
                "SIGNATURE",
                "SIGNATURE_ADULT",
                "NO_SIGNATURE_REQUIRED",
                "SPECIAL_HANDLING",
                "SPECIAL_EQUIPMENT",
                "TAILGATE_ORIGIN",
                "TAILGATE_DESTINATION",
                "SINGLE_SHIPMENT",
                "LIMITED_ACCESS",
                "MILITARY_BASE_DELIVERY",
                "HOMELAND_SECURITY",
                "INSIDE_DELIVERY",
                "RTS",
                "ABAN",
                "SATURDAY_DELIVERY"
              ]
            }
          },
          "shipDate": {
            "type": "string",
            "description": "Shipment date (Format: yyyy-MM-dd)",
            "format": "date-time"
          },
          "insuranceType": {
            "type": "string",
            "description": "Insurance type, defaults to CARRIER if not specified",
            "enum": [
              "SHIPTIME",
              "CARRIER"
            ]
          },
          "customsInvoice": {
            "$ref": "#/components/schemas/CustomsInvoiceModel"
          }
        },
        "description": "Rate request with shipment parameters. It, generally, should be exactly the same during rating and shipping to ensure quote and billing amounts consistency.\n\nPlease note that carrier APIs usually apply more stringent data validation when placing a shipment than during rating. For example, address field lengths have various restrictions depending on label design; phone number validity gets checked to ensure delivery, etc. These all may not be applied during rating for which sometimes only postal codes are enough.\n\nBut, to make sure the same rateRequest works well during rating and shipping, please, take care to fill out shipping details accurately and enter textual fields as short as possible especially when it comes to address details so it fits on a label. Keep in mind that label fields restrictions are different from carrier to carrier."
      },
      "ShipRequest": {
        "type": "object",
        "properties": {
          "rateRequest": {
            "$ref": "#/components/schemas/RateRequest"
          },
          "carrierId": {
            "type": "string",
            "description": "Carrier Id selected based on previous rating result"
          },
          "serviceId": {
            "type": "string",
            "description": "Service Id of the selected carrier"
          },
          "quoteId": {
            "type": "string",
            "description": "The quote Id referencing a selected quote from rating response. Will only be used if no other parameters are present in request, otherwise is ignored."
          },
          "pickupDetail": {
            "$ref": "#/components/schemas/PickupDetailModel"
          },
          "ref1": {
            "type": "string",
            "description": "Reference 1"
          },
          "ref2": {
            "type": "string",
            "description": "Reference 2"
          }
        },
        "description": "Represents request to create a shipment. If all properties are set, the rateRequest, carrierId, and serviceId will be used and quoteId is discarded.Note that shipment charge will be based on the rateRequest and the shipment billing amount may differ from the amount of the quote. This can happen if rateRequest is altered, that is not the same as during rating.\n\nIn order to ensure billing consistency between rating and shipping one can use the quoteId only in shipRequest, without setting other properties.\n\nNote, that the quoteId is only valid for about 15 min since rating."
      },
      "InvoiceItemModel": {
        "required": [
          "code",
          "description",
          "origin",
          "provinceOrState",
          "quantity",
          "unitPrice"
        ],
        "type": "object",
        "properties": {
          "quantity": {
            "type": "integer",
            "description": "Quantity of the item",
            "format": "int32"
          },
          "code": {
            "maxLength": 10,
            "minLength": 6,
            "type": "string",
            "description": "HS Code - 6 characters or 10 characters"
          },
          "description": {
            "maxLength": 50,
            "minLength": 0,
            "type": "string",
            "description": "Description of the item. Maximum: 50 characters"
          },
          "origin": {
            "type": "string",
            "description": "Country of Origin"
          },
          "provinceOrState": {
            "type": "string",
            "description": "Province of Origin. Required by some carriers (ex. CanadaPost) when country of origin is CA"
          },
          "unitPrice": {
            "type": "integer",
            "description": "Unit price of one line item represented in cents",
            "format": "int32"
          },
          "weight": {
            "type": "number",
            "description": "Weight of one line item. Note that most carriers require that total invoice weight of invoice items does not exceed the weight of shipment.",
            "format": "float"
          }
        },
        "description": "Represents the items in a commercial/customs invoice. Note that many carriers require the total insured value amount of a shipment to not exceed total customs value amount"
      },
      "ShipResponse": {
        "type": "object",
        "properties": {
          "shipId": {
            "type": "integer",
            "description": "Shipment Id",
            "format": "int32"
          },
          "trackingNumbers": {
            "type": "array",
            "description": "List of Tracking Numbers",
            "items": {
              "type": "string"
            }
          },
          "labelUrl": {
            "type": "string",
            "description": "Label URL"
          },
          "invoiceUrl": {
            "type": "string",
            "description": "Invoice URL"
          },
          "carrierTrackingUrl": {
            "type": "string",
            "description": "Carrier Tracking URL"
          },
          "pickupConfirmation": {
            "type": "string",
            "description": "Pickup Confirmation"
          },
          "messages": {
            "type": "array",
            "description": "Messages",
            "items": {
              "type": "string"
            }
          }
        },
        "description": "Represents results of request to create a shipment"
      },
...
    }
  },
...
}